upgrade webpack config files to ECMA script
