+++ packages/stacks-docs/webpack.config.mjs
@@ -1,13 +1,25 @@
-const fs = require("fs");
-const path = require("path");
-const { tsRule, lessRule, miniCssPlugin, commonResolve, commonDevServer } = require("../../webpack-common");
+import path from "path";
+import fs from "fs";
+import {
+    tsRule,
+    lessRule,
+    miniCssPlugin,
+    commonResolve,
+} from "../stacks-classic/webpack.config.mjs";
 
-module.exports = (_, argv) => {
-    const isProd = argv.mode === "production"
+import { fileURLToPath } from 'node:url';
+import { dirname } from 'node:path';
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
+
+export default (_, argv) => {
+    const isProd = argv.mode === "production";
 
     // load each entry.*.js file in assets/js as its own bundle
-    const entries = fs.readdirSync(path.resolve(__dirname, "assets/js/"))
-        .filter(f => f.startsWith("entry."))
+    const entries = fs
+        .readdirSync(path.resolve(__dirname, "assets/js/"))
+        .filter((f) => f.startsWith("entry."))
         .reduce((p, n) => {
             // { "entry.file": "path/to/entry.file.js" }
             p[n.slice(0, -3)] = path.resolve(__dirname, "assets/js/", n);
@@ -19,7 +31,7 @@ module.exports = (_, argv) => {
         devtool: isProd ? false : "inline-source-map",
         entry: {
             docs: path.resolve(__dirname, "assets/js/index.ts"),
-            ...entries
+            ...entries,
         },
         output: {
             filename: "[name].js",
@@ -32,10 +44,13 @@ module.exports = (_, argv) => {
                 lessRule(isProd),
             ],
         },
-        plugins: [
-            miniCssPlugin(),
-        ],
+        plugins: [miniCssPlugin()],
         resolve: commonResolve,
-        devServer: commonDevServer,
+        devServer: {
+            webSocketURL: {
+                // 11ty/browsersync steal the default port (8080), so set it to something else
+                port: 8081,
+            },
+        },
     };
 };
